# Improve ndimage convolve/correlate