This PR improves performance of `convolve` and `correlate` in `cupyx.scipy.ndimage.filter.py`.

I was investing why the performance of `convolve` and `correlate` was slower than expected, and it seems one of the causes is that raw array access via `CArray` is slow.. The issue will be addressed with this PR.

Performance (convolution, fp32)

[sec]  | scipy convolve | cupyx convolve (current) | cupyx convolve (this PR)
-----  | ---------------- | --------------------------- | ---------------------------
x.shape = (4000, 4000), w.shape = (3,3) | 0.225978 | 0.002195 | 0.000670
x.shape = (4000, 4000), w.shape = (5,5) | 0.390014 | 0.005525 | 0.001229
x.shape = (4000, 4000), w.shape = (7,7) | 0.761595 | 0.011325 | 0.002043
x.shape = (300, 300, 300), w.shape = (3, 3, 3) | 0.702484 | 0.028360 | 0.002025

* Quadro GV100
* CUDA 10.2