@grlee77 2020-03-10
That is a very nice speedup for such a small change!

Not directly related to this PR, but for cases when the kernel is small enough to fit, do you think it would help to copy it to shared memory? I suppose with a `RawModule` you could also keep the kernel in global/constant memory via `get_global`, but I don't think there is a way to do the equivalent with an `ElementwiseKernel`.