@leofang 2020-03-10
This is a very good catch, @anaruse! Do you think the `CArray` class is too abstract to provide good performance? I've had a hard time understanding it...