@coderforlife 2020-03-10
Note that I currently have a similar solution but then also generalizes to all of ndimage.filters. I too did notice the slow CArray lookups and then when using custom handling of indices it was way faster.