@coderforlife 2020-03-10
Your current version does always force the input array to be C contiguous and if it isn't copies the entire array. With a little extra work as done in mine (which I am going to submit a PR for in just a bit) it doesn't require non-contiguous arrays to be copied first.