@grlee77 2020-03-11
> Your current version does always force the input array to be C contiguous and if it isn't copies the entire array. With a little extra work as done in mine (which I am going to submit a PR for in just a bit) it doesn't require non-contiguous arrays to be copied first.

I agree that this is a limitation, but it could potentially be worth always enforcing contiguity for performance? We would have to benchmark on various contiguous and non-contiguous inputs to be sure. In memory-limited scenarios, avoiding the copy would obviously be preferable, though.

In any case, contiguity was already being enforced prior to this PR, so at least it is not a new requirement:
https://github.com/cupy/cupy/blob/05c8194dc9955f5fad2fbd8fb538389decca1756/cupyx/scipy/ndimage/filters.py#L106-L107