@coderforlife 2020-03-11
@grlee77 I did some very basic timing tests on my machine and having the non-contiguous helping of the input image (but not the weights, I still always make that contiguous since it should be relatively small) did not make a time difference with my code as in the other PR. I didn't actually test with any non-contiguous array, but the lack of speed difference isn't surprising since the current code essentially multiples by the stride right now anyways:

https://github.com/cupy/cupy/blob/05c8194dc9955f5fad2fbd8fb538389decca1756/cupyx/scipy/ndimage/filters.py#L175

It just computes the `sx_j` variables (the strides) from the shape instead of just using `x.strides()[{j}]`.

Here are the times for some convolutions, first line is not assuming contiguous arrays and second assumes contiguous (but doesn't actually do the check). Sometimes assuming contiguous is slower.

(4000, 4000) ⊗ (3, 3)
1.32 ms ± 1.34 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
1.31 ms ± 6.88 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) # assuming contiguous

(4000, 4000) ⊗ (5, 5)
2.91 ms ± 6.05 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
2.95 ms ± 9.63 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) # assuming contiguous

(4000, 4000) ⊗ (7, 7)
5.15 ms ± 2.04 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
5.25 ms ± 6.88 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) # assuming contiguous

(300, 300, 300) ⊗ (3, 3, 3)
17.1 ms ± 2.87 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
17.1 ms ± 4.89 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) # assuming contiguous