@coderforlife 2020-03-12
@leofang The problem with `CArray` is that if you use a linear index it incurs several divisions/modulos to compute the new index (even if that index will end up being the same exact same index since it is a contiguous array). That means every single lookup or set has a fairly large unexpected overhead.

I understand why it is doing this - it is to protect `ElementwiseKernel` authors from having to worry about non-contiguous array access. But it would be nice if:

- The array was contiguous then `CArray` directly used the linear index (and no division/modulo/multiplication)
- The `ElementwiseKernel` author could optionally turn off linear index conversion (even for non-contiguous arrays)

I can't remember if the tuple-index lookup has the same division problem, but I suspect not since I think it should only need the multiplications.