@emcastillo 2020-03-12
> The array was contiguous then CArray directly used the linear index (and no division/modulo/multiplication)


This should be fairly easy to do