@emcastillo 2020-03-12
I will allow this PR to go to the master branch while I review #3184. This will allow having some performance benefits right now.

Jenkins, test this please